node stackwalk for Electron
